Update README.md
Spent hours trying to figure out why it was showing Microsoft settings and realized I was pulling the wrong repository.   Just want to help others that might run into this problem.

Also add in the documentation that it runs best on Python 2.7.   I had Python 2.6.6 and a lot of the functions like:
    google.oauth2 import service_account
    from google.cloud import speech
    from google.cloud.speech import enums
    from google.cloud.speech import types
did not work and then I installed Python3.7 and the following did not work:
payload.set_payload(email.base64MIME.encode(mp3conv[2]))

but on Python 2.7, everything worked correctly.

Thanks for creating this application.  It was very helpful.